Added comments, formated the tags and renamed React components to use the .jsx extension
