+++ src/App.jsx
@@ -1,16 +1,31 @@
-import logo from './logo.svg';
+/**
+ * Tip: 
+ * Rename your React components as .jsx instead of .js to see a nice atom/electron icon on VS Code
+ * and to quickly differenciate it from normal JavaScript files.
+ */
+
+import logo from './logo.svg'; // always remove unused stuff
 import './App.css';
 import Task from './components/Task';
 import Filters from './components/Filters';
 import {useState} from 'react';
 
 function App() {
+  // To many booleans for a small project, take a step back and check how to simplify it.
   const [showCompleted, setShowCompleted] = useState(false);
   const [toggle, setToggle] = useState(true);
   const [toggleText, setToggleText] = useState("View completed tasks");
 
+  // because we only have 2 criteria, we dont need 2 booleans to keep track of stuff. 
   const [byPrice, setByPrice] = useState(false);
   const [byName, setByName] = useState(false);
+
+  // 1. Perfect, using dummy data is a nice way to kickstart the project.
+  /**
+   * 2. The key with the value Math.random() is not neccesary. 
+   * If we use functional programming, we can use the index of the array as a key
+   * to sort the list.
+  */
 
   const [list, setList] = useState([
       {key:Math.random(), isFinished: true ,name: "C", price: 500},
@@ -27,6 +42,7 @@ const sortByName = (e) => {
   e.preventDefault();
   let sorted;
 
+  // You dont need a if/else here. This function sorts by name, thus, you can create a new list based on this directly.
   if (byName){
       sorted = list.sort((a, b) => {
       return a.name.localeCompare(b.name);
@@ -45,6 +61,7 @@ const sortByPrice = (e) => {
   e.preventDefault();
   let sorted;
 
+  // Same as the other function
   if (byPrice){
     sorted = [... list].sort((a, b) => {
       return a.price - b.price;
@@ -59,7 +76,7 @@ const sortByPrice = (e) => {
   setList(sorted);
 }
 
-
+// Good!
 const notCompleted = list.filter(item => item.isFinished === false).map((item) =>(
   <Task key={item.key} 
         changed={setToCompleted}
@@ -68,6 +85,7 @@ const notCompleted = list.filter(item => item.isFinished === false).map((item) =
         price={item.price}></Task>
 ))
 
+// Good as well
 const completed = list.filter(item => item.isFinished === true).map((item) =>(
   <Task key={item.key} 
         state={item.isFinished} 
@@ -85,6 +103,7 @@ function setToCompleted(){
   })
 }
 
+// I will explain latter how to clean this one up.
 function change(e){
   e.preventDefault();
   if (toggle){
